+++ attributes/default.rb
@@ -18,11 +18,13 @@
 #
 #
 default['chef-server']['version'] = nil
+default['chef-server']['package_source'] = nil
 
 # The Chef Server must have an API FQDN set.
 # Ref. http://docs.chef.io/install_server_pre.html#hostnames
 default['chef-server']['api_fqdn'] = node['fqdn']
 
+default['chef-server']['topology'] = 'standalone'
 default['chef-server']['addons'] = []
 
 #
@@ -35,18 +37,18 @@ default['chef-server']['addons'] = []
 #
 # In a recipe:
 #
-#     node.override['chef-server']['configuration']['nginx']['ssl_port'] = 4433
+#     node.override['chef-server']['configuration'] = <<-EOS
+#     nginx['ssl_port'] = 4433
+#     EOS
 #
 # In a role:
 #
 #     default_attributes(
 #       'chef-server' => {
-#         'configuration' => {
-#           'nginx' => {
-#             'ssl_port' => 4433
-#           }
-#         }
+#         'configuration' => <<-EOS
+#     nginx['ssl_port'] = 4433
+#     EOS
 #       }
 #     )
 #
-default['chef-server']['configuration'] = {}
+default['chef-server']['configuration'] = ''